+++ cli/cmd_slq.go
@@ -422,11 +422,11 @@ func addQueryCmdFlags(cmd *cobra.Command) {
 	addOptionFlag(cmd.Flags(), OptFormatDecimal)
 	panicOn(cmd.RegisterFlagCompletionFunc(
 		OptFormat.Flag().Name,
-		completeStrings(-1, stringz.Strings(format.All())...),
+		completeStrings(stringz.Strings(format.All())...),
 	))
 	panicOn(cmd.RegisterFlagCompletionFunc(
 		OptFormatDecimal.Flag().Name,
-		completeStrings(-1, "string", "number"),
+		completeStrings("string", "number"),
 	))
 	addResultFormatFlags(cmd)
 	cmd.MarkFlagsMutuallyExclusive(append(
@@ -446,7 +446,7 @@ func addQueryCmdFlags(cmd *cobra.Command) {
 		(&handleTableCompleter{onlySQL: true, handleRequired: true}).complete))
 
 	cmd.Flags().String(flag.ActiveSrc, "", flag.ActiveSrcUsage)
-	panicOn(cmd.RegisterFlagCompletionFunc(flag.ActiveSrc, completeHandle(0, false)))
+	panicOn(cmd.RegisterFlagCompletionFunc(flag.ActiveSrc, completeHandleFlag(false)))
 
 	cmd.Flags().String(flag.ActiveSchema, "", flag.ActiveSchemaUsage)
 	panicOn(cmd.RegisterFlagCompletionFunc(flag.ActiveSchema,
@@ -459,7 +459,7 @@ func addQueryCmdFlags(cmd *cobra.Command) {
 	addOptionFlag(cmd.Flags(), driver.OptIngestHeader)
 	addOptionFlag(cmd.Flags(), driver.OptIngestCache)
 	addOptionFlag(cmd.Flags(), csv.OptDelim)
-	panicOn(cmd.RegisterFlagCompletionFunc(csv.OptDelim.Key(), completeStrings(-1, csv.NamedDelims()...)))
+	panicOn(cmd.RegisterFlagCompletionFunc(csv.OptDelim.Key(), completeStrings(csv.NamedDelims()...)))
 	addOptionFlag(cmd.Flags(), csv.OptEmptyAsNull)
 }